Break apart tree logic a bit more and add new helper/test for detecting shared parent
